urcarcher-ops/requirements.txt
websockets
bs4
selenium==4.10.0
numpy==1.26.4
pandas==2.1.4
tensorflow==2.17.0
keras==3.4.1
scikit-learn==1.3.2
matplotlib
ray==2.35.0
python-dotenv

apache-airflow